+++ app/controllers/AuthController.php
@@ -4,10 +4,13 @@ namespace app\controllers;
 
 use Yii;
 use yii\filters\AccessControl;
+use yii\web\NotFoundHttpException;
 use yii\web\Response;
 use yii\filters\VerbFilter;
 use app\forms\LoginForm;
-use app\forms\ContactForm;
+use app\forms\PasswordRequestForm;
+use app\forms\PasswordUpdateForm;
+use app\models\User;
 
 class AuthController extends Controller
 {
@@ -44,7 +47,7 @@ class AuthController extends Controller
 	 */
 	public function actionLogin()
 	{
-		if ( ! Yii::$app->user->isGuest) {
+		if (! Yii::$app->user->isGuest) {
 			return $this->goHome();
 		}
 
@@ -69,6 +72,89 @@ class AuthController extends Controller
 		Yii::$app->user->logout();
 
 		return $this->goHome();
+	}
+
+	/**
+	 * Register action.
+	 *
+	 * @return string|Response
+	 */
+	public function actionRegister()
+	{
+		if (!Yii::$app->user->isGuest) {
+			return $this->goHome();
+		}
+
+		$model = new RegisterForm();
+
+		if ($model->load(Yii::$app->request->post()) && $model->register()) {
+			Yii::$app->session->addFlash('success', 'You have been successfully registered');
+
+			return $this->goBack();
+		}
+
+		return $this->render('register', [
+			'model' => $model,
+		]);
+	}
+
+	/**
+	 * Password request action
+	 *
+	 * @return string|Response
+	 */
+	public function actionPasswordRequest()
+	{
+		if (!Yii::$app->user->isGuest) {
+			return $this->goHome();
+		}
+
+		$model = new PasswordRequestForm();
+
+		if ($model->load(Yii::$app->request->post()) && $model->request()) {
+			Yii::$app->session->addFlash(
+				'success',
+				'If the email address is registered in the system, we would send a letter there shortly.'
+			);
+
+			return $this->goBack();
+		}
+
+		return $this->render('password-request', [
+			'model' => $model,
+		]);
+	}
+
+	/**
+	 * Password update action
+	 *
+	 * @param string $token Password reset token.
+	 *
+	 * @return string|Response
+	 * @throws NotFoundHttpException If token not found in DB.
+	 */
+	public function actionPasswordUpdate($token)
+	{
+		if (!Yii::$app->user->isGuest) {
+			return $this->goHome();
+		}
+
+		if (!User::isPasswordResetTokenValid($token)) {
+			throw new NotFoundHttpException('Page not found.');
+		}
+
+		$model = new PasswordUpdateForm();
+		$model->resetToken = $token;
+
+		if ($model->load(Yii::$app->request->post()) && $model->update()) {
+			Yii::$app->session->addFlash('success', 'Your password has been successfully updated!');
+
+			return $this->goBack();
+		}
+
+		return $this->render('password-update', [
+			'model' => $model,
+		]);
 	}
 
 }