+++ app/modules/admin/forms/UserForm.php
@@ -3,6 +3,7 @@
 namespace app\modules\admin\forms;
 
 use app\models\User;
+use Yii;
 
 class UserForm extends User
 {
@@ -10,12 +11,17 @@ class UserForm extends User
 	 * @var string
 	 */
 	public $password;
-
+	
 	/**
 	 * @var string
 	 */
-	public $password_repeat;
-
+	public $passwordRepeat;
+	
+	/**
+	 * @var string
+	 */
+	public $roles;
+	
 	/**
 	 * @inheritdoc
 	 * @return array
@@ -23,11 +29,20 @@ class UserForm extends User
 	public function rules()
 	{
 		return array_merge(parent::rules(), [
-			['password', 'compare'],
-			['password_repeat', 'safe'],
+			['password', 'safe'],
+			[
+				'passwordRepeat',
+				'required',
+				'when'       => function () {
+					return !empty($this->password);
+				},
+				'whenClient' => 'function() { return $.trim($("#userform-password").val()).length || false; }',
+			],
+			['passwordRepeat', 'compare', 'compareAttribute' => 'password'],
+			['roles', 'in', 'range' => array_keys(static::getRolesList()), 'allowArray' => true],
 		]);
 	}
-
+	
 	/**
 	 * @inheritdoc
 	 * @return bool
@@ -37,6 +52,37 @@ class UserForm extends User
 		if ($this->password) {
 			$this->setPassword($this->password);
 		}
+		
+		$this->updateUserRoles($this->roles);
+		
 		return parent::beforeSave($insert);
+	}
+	
+	/**
+	 * @inheritdoc
+	 */
+	public function afterFind()
+	{
+		$roles = array_keys(Yii::$app->authManager->getRolesByUser($this->id));
+		$roles = array_combine($roles, $roles);
+		$this->roles = $roles;
+		
+		parent::afterFind();
+	}
+	
+	/**
+	 * @param array|string $roles
+	 */
+	protected function updateUserRoles($roles)
+	{
+		if (empty($roles)) {
+			return;
+		}
+		
+		Yii::$app->authManager->revokeAll($this->id);
+		
+		foreach ($roles as $role) {
+			$this->assignRole($role);
+		}
 	}
 }